ORCH-0849 follow-up: web-safe .native/.web split for mingla-business Stripe imports
PR #115 CI flagged a real regression: I-PROPOSED-AE STRIPE_REACT_NATIVE_NATIVE_BOUNDARY_ONLY
(enforced by orch-0778-web-stripe-native-import-gate.mjs) blocked the merge because
ORCH-0849 imported @stripe/stripe-react-native into mingla-business routes that resolve
on web (Vercel). mingla-business ships on web; pulling Stripe React Native into the web
bundle breaks `expo export --platform web` (native-only codegenNativeComponent).

Fix preserves both invariants — restores PaymentSheet parity (ORCH-0849) AND keeps the
web bundle Stripe-free (I-PROPOSED-AE / ORCH-0778) — via platform extension files:

- mingla-business/src/payments/nativeCheckoutFlow.native.ts: real impl (moved from
  nativeCheckoutFlow.ts; unchanged semantics).
- mingla-business/src/payments/nativeCheckoutFlow.ts: web-safe stub. Returns
  {outcome: "failed"} if ever called; not invoked on web because payment.tsx
  branches on Platform.OS === "web" before calling the hook.
- mingla-business/src/payments/StripeProviderWrapper.native.tsx: real wrapper that
  mounts <StripeNativeProvider> with the business merchantIdentifier + urlScheme.
- mingla-business/src/payments/StripeProviderWrapper.tsx: passthrough Fragment (web).
- mingla-business/app/_layout.tsx: imports the local wrapper instead of
  @mingla/payments-native directly, so web bundle doesn't pull Stripe transitively.

CI gate updates:
- orch-0778-web-stripe-native-import-gate.mjs: allowlist updated from empty
  (ORCH-0839-B hosted-only era) to include nativeCheckoutFlow.native.ts. Comment
  rewritten to reflect ORCH-0849 reversal while keeping the invariant active.
- i-stripe-paymentsheet-parity.mjs: business-side constants point at the .native
  files so R-5..R-8 inspect the real impl, not the stubs.

Test path:
- native_checkout_flow_parity.test.ts: file-path strings updated to .native.ts.

Local verification (clean tree):
  node .github/scripts/strict-grep/orch-0778-web-stripe-native-import-gate.mjs → PASS
  node .github/scripts/strict-grep/i-stripe-paymentsheet-parity.mjs            → PASS (8/8)
  node .github/scripts/strict-grep/i-stripe-pm-method-allowlist.mjs            → PASS (6/6)
